FGCObjectScopeGuard to avoid GC after ImportedSoundWave is created; int64 instead of int32 in Flac, MP3 and WAV transcoders

+++ Source/RuntimeAudioImporter/Private/Transcoders/FlacTranscoder.cpp
@@ -48,7 +48,7 @@ bool FlacTranscoder::Decode(FEncodedAudioStruct&& EncodedData, FDecodedAudioStru
 	DecodedData.PCMInfo.PCMNumOfFrames = drflac_read_pcm_frames_f32(FLAC_Decoder, FLAC_Decoder->totalPCMFrameCount, TempPCMData);
 
 	// Getting PCM data size
-	const int32 TempPCMDataSize = static_cast<int32>(DecodedData.PCMInfo.PCMNumOfFrames * FLAC_Decoder->channels * sizeof(float));
+	const int64 TempPCMDataSize = static_cast<int64>(DecodedData.PCMInfo.PCMNumOfFrames * FLAC_Decoder->channels * sizeof(float));
 
 	DecodedData.PCMInfo.PCMData = FBulkDataBuffer<float>(TempPCMData, TempPCMDataSize);
